- Removed the various workarounds associated with <unistd.h>, required   for building code that references the networking API. It is now possible,   provided you build your code with an existing TCP/IP API header file   set (such as from the "Roadshow" SDK) without running into data type   or function prototype declaration conflicts. I have added local versions   of <arpa/inet.h>, <netinet/in.h>, <netdb.h>, <sys/filio.h>, <sys/ioccom.h>,   <sys/ioctl.h>, <sys/select.h> and <sys/socket.h> and updated <unistd.h>,   <pwd.h> and <grp.h> accordingly. Note that some of the function prototypes   will conflict with the bsdsocket.library/usergroup.library API, which is   why you can disable the declarations through the __NO_NET_API preprocessor   symbol: if defined, none of the function prototypes will be defined.
git-svn-id: file:///Users/olsen/Code/migration-svn-zu-git/logical-line-staging/clib2/trunk@15036 87f5fb63-7c3d-0410-a384-fd976d0f7a62

+++ library/socket_gethostname.c
@@ -1,5 +1,5 @@
 /*
- * $Id: socket_gethostname.c,v 1.4 2005-04-24 08:46:37 obarthel Exp $
+ * $Id: socket_gethostname.c,v 1.5 2005-10-09 12:32:18 obarthel Exp $
  *
  * :ts=4
  *
@@ -48,7 +48,7 @@
 /****************************************************************************/
 
 int
-gethostname(const char *hostname,int size)
+gethostname(const char *hostname,size_t size)
 {
 	int result = ERROR;
